refact: split db manager, postgres
