+++ backend/pkg/dbmanager/postgres_driver.go
@@ -3,7 +3,6 @@ package dbmanager
 import (
 	"context"
 	"database/sql"
-	"encoding/json"
 	"fmt"
 	"log"
 	"neobase-ai/internal/apis/dtos"
@@ -21,98 +20,6 @@ type PostgresDriver struct{}
 
 func NewPostgresDriver() DatabaseDriver {
 	return &PostgresDriver{}
-}
-
-// Add these constants
-const (
-	QueryTypeUnknown = "UNKNOWN"
-	QueryTypeDDL     = "DDL"
-	QueryTypeDML     = "DML"
-	QueryTypeSelect  = "SELECT"
-)
-
-// Add these types for PostgreSQL schema tracking
-type PostgresSchema struct {
-	Tables      map[string]PostgresTable
-	Indexes     map[string][]PostgresIndex
-	Views       map[string]PostgresView
-	Sequences   map[string]PostgresSequence     // For auto-increment/serial
-	Constraints map[string][]PostgresConstraint // Table constraints (CHECK, UNIQUE, etc.)
-	Enums       map[string]PostgresEnum         // Custom enum types
-}
-
-type PostgresTable struct {
-	Name        string
-	Columns     map[string]PostgresColumn
-	Indexes     map[string]PostgresIndex
-	PrimaryKey  []string
-	ForeignKeys map[string]PostgresForeignKey
-	RowCount    int64
-}
-
-type PostgresColumn struct {
-	Name         string
-	Type         string
-	IsNullable   bool
-	DefaultValue string
-	Comment      string
-}
-
-type PostgresIndex struct {
-	Name      string
-	Columns   []string
-	IsUnique  bool
-	TableName string
-}
-
-type PostgresView struct {
-	Name       string
-	Definition string
-}
-
-type PostgresForeignKey struct {
-	Name      string
-	Column    string
-	RefTable  string
-	RefColumn string
-	OnDelete  string
-	OnUpdate  string
-}
-
-// Add new types for additional schema elements
-type PostgresSequence struct {
-	Name       string
-	StartValue int64
-	Increment  int64
-	MinValue   int64
-	MaxValue   int64
-	CacheSize  int64
-	IsCycled   bool
-}
-
-type PostgresConstraint struct {
-	Name       string
-	Type       string // CHECK, UNIQUE, etc.
-	TableName  string
-	Definition string
-	Columns    []string
-}
-
-type PostgresEnum struct {
-	Name   string
-	Values []string
-	Schema string
-}
-
-// Add a conversion method for PostgresColumn
-func (pc PostgresColumn) toColumnInfo() ColumnInfo {
-	return ColumnInfo{
-		Name:         pc.Name,
-		Type:         pc.Type,
-		IsNullable:   pc.IsNullable,
-		DefaultValue: pc.DefaultValue,
-		Comment:      pc.Comment,
-	}
 }
 
 func (d *PostgresDriver) Connect(config ConnectionConfig) (*Connection, error) {
@@ -332,21 +239,6 @@ func (d *PostgresDriver) ExecuteQuery(ctx context.Context, conn *Connection, que
 	return result
 }
 
-// Helper function to split SQL statements
-func splitStatements(query string) []string {
-	// Basic statement splitting - can be enhanced for more complex cases
-	statements := strings.Split(query, ";")
-
-	// Clean up statements
-	var result []string
-	for _, stmt := range statements {
-		if stmt = strings.TrimSpace(stmt); stmt != "" {
-			result = append(result, stmt)
-		}
-	}
-	return result
-}
-
 func (d *PostgresDriver) BeginTx(ctx context.Context, conn *Connection) Transaction {
 	log.Printf("PostgreSQL/YugabyteDB Driver -> BeginTx -> Starting transaction")
 
@@ -372,222 +264,6 @@ func (d *PostgresDriver) BeginTx(ctx context.Context, conn *Connection) Transact
 		tx:   tx,
 		conn: conn,
 	}
-}
-
-// Update PostgresTransaction to handle schema updates
-type PostgresTransaction struct {
-	tx   *sql.Tx
-	conn *Connection // Add connection reference
-}
-
-func (tx *PostgresTransaction) ExecuteQuery(ctx context.Context, conn *Connection, query string, queryType string, findCount bool) *QueryExecutionResult {
-	startTime := time.Now()
-
-	// Split into individual statements
-	statements := splitStatements(query)
-	log.Printf("PostgreSQL Transaction -> ExecuteQuery -> Statements: %v", statements)
-
-	var lastResult sql.Result
-	var rows *sql.Rows
-	var err error
-
-	for _, stmt := range statements {
-		stmt = strings.TrimSpace(stmt)
-		if stmt == "" {
-			continue
-		}
-
-		// For SELECT queries
-		if strings.HasPrefix(strings.ToUpper(stmt), "SELECT") {
-			rows, err = tx.tx.QueryContext(ctx, stmt)
-			if err != nil {
-				return &QueryExecutionResult{
-					Error: &dtos.QueryError{
-						Code:    "QUERY_EXECUTION_FAILED",
-						Message: err.Error(),
-						Details: fmt.Sprintf("Failed to execute SELECT: %s", stmt),
-					},
-				}
-			}
-		} else {
-			// For non-SELECT queries
-			lastResult, err = tx.tx.ExecContext(ctx, stmt)
-			if err != nil {
-				return &QueryExecutionResult{
-					Error: &dtos.QueryError{
-						Code:    "QUERY_EXECUTION_FAILED",
-						Message: err.Error(),
-						Details: fmt.Sprintf("Failed to execute %s: %s", queryType, stmt),
-					},
-				}
-			}
-
-			// Check for specific PostgreSQL errors
-			if strings.Contains(strings.ToUpper(stmt), "DROP TABLE") {
-				// Extract table name
-				tableName := extractTableName(stmt)
-				log.Printf("PostgresDriver -> ExecuteQuery -> Checking existence of table: %s", tableName)
-
-				// Check if table exists before dropping
-				var exists bool
-				checkStmt := `SELECT EXISTS (SELECT 1 FROM information_schema.tables WHERE table_schema='public' AND table_name=$1)`
-
-				err = tx.tx.QueryRow(checkStmt, tableName).Scan(&exists)
-				if err != nil {
-					log.Printf("PostgresDriver -> ExecuteQuery -> Error checking table existence: %v", err)
-					return &QueryExecutionResult{
-						Error: &dtos.QueryError{
-							Code:    "TABLE_NOT_FOUND",
-							Message: err.Error(),
-							Details: "Cannot drop a table that doesn't exist",
-						},
-					}
-				}
-
-				log.Printf("PostgresDriver -> ExecuteQuery -> Table '%s' exists, proceeding with DROP", tableName)
-			}
-		}
-	}
-
-	// Process results
-	result := &QueryExecutionResult{
-		ExecutionTime: int(time.Since(startTime).Milliseconds()),
-	}
-
-	if rows != nil {
-		defer rows.Close()
-		results, err := processRows(rows, startTime)
-		if err != nil {
-			return &QueryExecutionResult{
-				ExecutionTime: int(time.Since(startTime).Milliseconds()),
-				Error: &dtos.QueryError{
-					Code:    "RESULT_PROCESSING_FAILED",
-					Message: err.Error(),
-					Details: "Failed to process query results",
-				},
-			}
-		}
-		result.Result = map[string]interface{}{
-			"results": results,
-		}
-	} else if lastResult != nil {
-		rowsAffected, _ := lastResult.RowsAffected()
-		if rowsAffected > 0 {
-			result.Result = map[string]interface{}{
-				"rowsAffected": rowsAffected,
-				"message":      fmt.Sprintf("%d row(s) affected", rowsAffected),
-			}
-		} else {
-			result.Result = map[string]interface{}{
-				"message": "Query performed successfully",
-			}
-		}
-	}
-
-	resultJSON, err := json.Marshal(result.Result)
-	if err != nil {
-		return &QueryExecutionResult{
-			ExecutionTime: int(time.Since(startTime).Milliseconds()),
-			Error: &dtos.QueryError{
-				Code:    "JSON_MARSHAL_FAILED",
-				Message: err.Error(),
-				Details: "Failed to marshal query results",
-			},
-		}
-	}
-	result.ResultJSON = string(resultJSON)
-
-	return result
-}
-
-// Fix the extractTableName function to properly handle table names
-func extractTableName(stmt string) string {
-	// Preserve original case for the extraction
-	originalStmt := strings.TrimSpace(stmt)
-	upperStmt := strings.ToUpper(originalStmt)
-
-	var tableName string
-
-	// Handle different DROP TABLE variations
-	if strings.HasPrefix(upperStmt, "DROP TABLE IF EXISTS") {
-		tableName = strings.TrimSpace(originalStmt[len("DROP TABLE IF EXISTS"):])
-	} else if strings.HasPrefix(upperStmt, "DROP TABLE") {
-		tableName = strings.TrimSpace(originalStmt[len("DROP TABLE"):])
-	} else {
-		return ""
-	}
-
-	// Handle schema prefixes, quotes, and trailing characters
-	tableName = strings.Split(tableName, " ")[0] // Get first word
-	tableName = strings.Trim(tableName, "\"`;")  // Remove quotes and semicolons
-
-	// Handle schema prefixes like "public."
-	if strings.Contains(tableName, ".") {
-		parts := strings.Split(tableName, ".")
-		tableName = parts[len(parts)-1] // Get the last part after the dot
-	}
-
-	log.Printf("PostgresDriver -> extractTableName -> Extracted table name '%s' from statement: %s",
-		tableName, originalStmt)
-
-	return tableName
-}
-
-func (t *PostgresTransaction) Commit() error {
-	log.Printf("PostgreSQL Transaction -> Commit -> Committing transaction")
-	return t.tx.Commit()
-}
-
-func (t *PostgresTransaction) Rollback() error {
-	log.Printf("PostgreSQL Transaction -> Rollback -> Rolling back transaction")
-	return t.tx.Rollback()
-}
-
-// Update the processRows function signature to return results and error
-func processRows(rows *sql.Rows, startTime time.Time) ([]map[string]interface{}, error) {
-	columns, err := rows.Columns()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get columns: %v", err)
-	}
-
-	results := make([]map[string]interface{}, 0)
-	values := make([]interface{}, len(columns))
-	scanArgs := make([]interface{}, len(columns))
-
-	for i := range values {
-		scanArgs[i] = &values[i]
-	}
-
-	for rows.Next() {
-		err := rows.Scan(scanArgs...)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan row: %v", err)
-		}
-
-		row := make(map[string]interface{})
-		for i, col := range columns {
-			val := values[i]
-			if val == nil {
-				row[col] = nil
-				continue
-			}
-
-			// Handle different types
-			switch v := val.(type) {
-			case []byte:
-				row[col] = string(v)
-			default:
-				row[col] = v
-			}
-		}
-		results = append(results, row)
-	}
-
-	if err = rows.Err(); err != nil {
-		return nil, fmt.Errorf("error iterating rows: %v", err)
-	}
-
-	return results, nil
 }
 
 // Improve the GetSchema method to properly detect all tables
